streaming-etl/services/transformer/pyproject.toml
[tool.poetry]
name = "transformer"
version = "0.1.0"
description = ""
authors = ["Beaglefoot <ch-e-st@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
aiokafka = "^0.7.2"
pydantic = "^1.9.0"
python-schema-registry-client = "^2.4.0"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
datamodel-code-generator = {extras = ["http"], version = "^0.12.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
